made things working
```
NEXTDNS_PROFILE=... NEXTDNS_API_KEY=... go test ./...
?       github.com/raylas/nextdns-exporter      [no test files]
ok      github.com/raylas/nextdns-exporter/internal/api (cached)
ok      github.com/raylas/nextdns-exporter/internal/util        0.150s
```
